refactor: complete layered architecture refactor
