+++ backend/src/application/get-videos-for-user.js
@@ -1,6 +1,6 @@
-const DynamoFactory = require("./factories/DynamoFactory");
-const extractItemsFromDynamoResults = require("./extract-items-from-dynamo-results");
-const getTagsAvailableToUser = require("../application/get-tags-available-to-user");
+const DynamoFactory = require("../utility/factories/DynamoFactory");
+const extractItemsFromDynamoResults = require("../utility/extract-items-from-dynamo-results");
+const getTagsAvailableToUser = require("./get-tags-available-to-user");
 
 module.exports = async userHash => {
 	const videos = await getVideos(userHash);
