+++ backend/src/application/thumbnail-generation/does-file-exist.js
