+++ backend/src/application/thumbnail-generation/generate-tmp-file-path.js
