+++ backend/src/application/thumbnail-generation/make-thumbnail.js
