+++ backend/src/application/validate-auth-token.js
