+++ backend/src/persistence/get-list-of-users.js
@@ -1,5 +1,5 @@
-const extractItemsFromDynamoResults = require("../lib/extract-items-from-dynamo-results");
-const DynamoFactory = require("../lib/factories/DynamoFactory");
+const extractItemsFromDynamoResults = require("../utility/extract-items-from-dynamo-results");
+const DynamoFactory = require("../utility/factories/DynamoFactory");
 
 module.exports = async () => {
 	const users = await getAllUsers();
