+++ backend/src/utility/extract-items-from-dynamo-results.js
