+++ backend/src/utility/factories/DynamoFactory.js
