+++ backend/src/utility/factories/S3Factory.js
